requirements.txt
matplotlib
numpy
pandas
scikit-learn
seaborn
tensorflow-gpu
